requirements.txt
iniconfig==2.1.0
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
numpy==2.3.3
packaging==25.0
pandas==2.3.3
platformdirs==4.4.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rich==14.1.0
six==1.17.0
textual==6.1.0
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
